dodana ocena časovne zahtevnosti in preverjanje kdo lahko spreminja podatke

+++ scrum/views.py
@@ -692,6 +692,9 @@ def new_user_story(request, project_name):
     if request.method == "POST":
         form = UserStoryForm(request.POST)
         if form.is_valid():
+            if not methodology_manager:
+                if form.cleaned_data['original_estimate'] is not None:
+                    messages.error(request, "Original estimate can be provided only by Methology Manager")
             if methodology_manager or product_owner:
                 instance = form.save() 
                 messages.success(request, f"User story \"{instance.name}\" created!!")
@@ -711,19 +714,28 @@ def new_user_story(request, project_name):
         return render(request,'new_user_story.html',context=context)
     
 def edit_user_story(request, project_name, id):
+    context = get_context(request)
     project = Project.objects.get(name = project_name)
     user_story = UserStory.objects.get(id=id)
     form = UserStoryForm(instance=user_story, initial={'project':project})
+    methodology_manager = AssignedRole.objects.filter(project = project, user=context['id'],role = 'methodology_manager')
+    product_owner = AssignedRole.objects.filter(project = project, user=context['id'],role = 'product_owner')
     if request.method == "POST":
         form = UserStoryForm(request.POST, instance=user_story)
         if form.is_valid():
-            form.save()
-            messages.success(request, f"User story \"{user_story.name}\" updated!!")
-            return redirect('project_name', project_name=project_name)
+            if not methodology_manager:
+                if form.cleaned_data['original_estimate'] != user_story.original_estimate:
+                    messages.error(request, "Original estimate can be provided only by Methology Manager")
+            if methodology_manager or product_owner:
+                form.save() 
+                messages.success(request, f"User story \"{user_story.name}\" updated!!")
+                return redirect('project_name', project_name=project_name)
+            else:
+                messages.error(request, "User story can be updated only by Product Owner or Methology Manager")
+                return redirect(request.path)
         else:
             messages.error(request, form.errors)
             return redirect(request.path)
-    context = get_context(request)
     context['form'] = form
     context['project'] = project
     return render(request,'new_user_story.html',context=context)
@@ -733,14 +745,20 @@ def delete_user_story(request, project_name, id):
     user_story = UserStory.objects.get(id=id)
     user_story_name = user_story.name
     project = Project.objects.get(name = project_name)
+    methodology_manager = AssignedRole.objects.filter(project = project, user=context['id'],role = 'methodology_manager')
+    product_owner = AssignedRole.objects.filter(project = project, user=context['id'],role = 'product_owner')
     print(user_story.sprint)
     if request.method == "POST":
-        if user_story.sprint is None:
-            user_story.delete()
-            messages.success(request, f"User story \"{user_story_name}\" deleted successfully!!")
-            return redirect('project_name', project_name=project_name)
+        if methodology_manager or product_owner:
+            if user_story.sprint is None:
+                user_story.delete()
+                messages.success(request, f"User story \"{user_story_name}\" deleted successfully!!")
+                return redirect('project_name', project_name=project_name)
+            else:
+                messages.error(request, f"User story \"{user_story.name}\" can't be deleted, because it is already in current sprint!!")
+                return redirect('project_name', project_name=project_name)
         else:
-            messages.error(request, f"User story \"{user_story.name}\" can't be deleted, because it is already in current sprint!!")
+            messages.error(request, "User story can be deleted only by Product Owner or Methology Manager")
             return redirect('project_name', project_name=project_name)
     context["user_story_id"] = user_story.id
     context["user_story_name"] = user_story.name
